Fix value range errors and warnings
what:
When we switch course code with a different course number
set, we implicitly update form for the course numbers.
Because we arent calling change handler, we dont update its
value. This results in the form keeping it and causing warnings as
well as unwanted behaviour

how: check for course number form value to be withing a given range
and if not reset it to nothing.

jira: FRON-45
jira: FRON-49